Summary
Delicious baking is a website to prepare several bakery items and vegetarian snack/meals recipes. All the baking recipes that are in the application are presented in a simple step-by-step instruction method.
We offer a lot of recipes from various categories:

    - Cakes
    - Cupcakes
    - Cookies
    - Desserts
    - Savoury
    - Breads
    - No Bakes
    - Kids No-bake: designed for kids to make by themselves with adult supervision.

We are also currently working on recipes for Dairy-free, pastires in our Featured Recipes Section.